feat: ajout des insertion du fichier extrait.yml
