--- composer.json
@@ -1,5 +1,0 @@
-{
-    "require": {
-        "symfony/yaml": "^6.4"
-    }
-}
